+++ web_flask/api/v1/stories.py
@@ -3,13 +3,14 @@
 
 """
 
-import json
+import web_flask.api.v1.services.auth_provider as auth
 from sqlalchemy.orm import joinedload
-from flask import request, jsonify, abort, url_for
-from flask_login import current_user, login_required
+from flask import request, jsonify, abort
 from web_flask.api.v1 import views
 from web_flask.api.v1.helper_func import create_uri, check_for_valid_json
-from web_flask.api.v1.helper_func import custom_login_required
+from web_flask.api.v1.services.auth_guard import auth_guard
+from web_flask.api.v1.services.data_service import get_story_data
+from web_flask.api.v1.services.data_service import get_comment_data
 from models.story import Story
 from models.user import User
 from models.comment import Comment
@@ -24,7 +25,7 @@ from web_flask.api.v1 import storage
     strict_slashes=False,
     methods=['GET', 'POST']
 )
-#@login_required
+@auth_guard
 def stories():
     """ Get all stories or creates a story """
 
@@ -32,7 +33,7 @@ def stories():
         # checks for valid json
         try:
             story_json = request.get_json()
-            check_for_valid_json(story_json, ['title', 'text', 'user_id'])
+            check_for_valid_json(story_json, ['title', 'text'])
 
         except Exception:
             return jsonify({"Error": 'not a valid json'}), 400
@@ -42,71 +43,57 @@ def stories():
                 story = Story(
                         title=story_json['title'],
                         text=story_json['text'],
-                        user_id=story_json['user_id']
+                        user_id=auth.current_user.id
                     )
             elif isinstance(story_json, list):
                 story = Story(
                     title=story_json[0],
                     text=story_json[1],
-                    user_id=story_json[2]
+                    user_id=auth.current_user.id
                 )
             storage.new(story)
             storage.save()
-            print(story)
-            return jsonify(story.to_dict()), 201
+            return jsonify(get_story_data(story)), 201
 
     stories = storage.all(Story)
 
     return jsonify(
-        [create_uri(story.to_dict(), 'get_story')
+        [create_uri(get_story_data(story), 'get_story')
          for story in stories.values()]
     ), 200
 
 
 @views.route(
-        '/users/<string:user_id>/following_stories/',
+        '/users/following_stories/',
         methods=['GET'],
         strict_slashes=False
 )
-#login_required
-def following_stories(user_id=None):
+@auth_guard
+def following_stories():
     """ gets the stories from all the users self is following
         and own stories
 
     """
-    user = storage.get(User, user_id)
-    if user is None:
+    if not auth.current_user:
         abort(404)
 
     page = request.args.get('page', 1, type=int)
     per_page = request.args.get('per_page', 10, type=int)
     
     # query = storage._session.query(Story).where(Story.user_id == user.id)
-    query = user.following_stories
+    query = auth.current_user.following_stories
     
     pagination = Story.paginate(query, page, per_page)
 
     stories = []
     for story in pagination['items']:
-        story_dictionary = story.to_dict()
-        story_dictionary['liked'] = user.liked_story(story.id)
-        story_dictionary['bookmarked'] = user.bookmarked_story(story.id)
-        story_dictionary['user_is_following_writer'] = user.is_following(story.writer)
-
-        stories.append(story_dictionary)
+        stories.append(get_story_data(story))
 
     if stories == []:
-        _stories = [ story.to_dict() for story in storage.all(Story).values() ]
         stories = []
-        for story in _stories:
-            try:
-                story['liked'] = user.liked_story(story['id'])
-                story['bookmarked'] = user.bookmarked_story(story['id'])
-                writer = storage.get(User, story['writer']['id'])
-                story['user_is_following_writer'] = user.is_following(writer)
-                stories.append(story)
-            except Exception:
-                pass
+        for story in storage.all(Story).values():
+            stories.append(get_story_data(story))
+            
         pagination = Story.paginate_list(stories)
         stories = pagination['items']
     
@@ -126,7 +113,7 @@ def following_stories(user_id=None):
     strict_slashes=False,
     methods=['GET', 'POST']
 )
-#login_required
+@auth_guard
 def limit_stories(n=None):
     """ Limits the number of stories to get
 
@@ -151,7 +138,7 @@ def limit_stories(n=None):
 
     return jsonify(
         [
-            create_uri(story.to_dict(), 'get_story')
+            create_uri(get_story_data(story), 'get_story')
             for story in limited_stories.values()
         ]
     ), 200
@@ -159,53 +146,71 @@ def limit_stories(n=None):
 
 @views.route(
     '/stories/<string:story_id>/',
-    methods=['GET', 'PUT'],
+    methods=['GET'],
     strict_slashes=False
 )
-#login_required
+@auth_guard
 def get_story(story_id=None):
     """ Gets a specific story or update an existing one """
     story = storage.get(Story, story_id)
     if story is None:
         abort(404)
 
-    if request.method == 'PUT':
-        try:
-            story_json = request.get_json()
-            check_for_valid_json(story_json, ['title', 'text'])  # removed check for user_id
+    # data
+    return jsonify(get_story_data(story)), 200
 
-        except Exception:
-            return jsonify({"Error": 'not a valid json'}), 400
 
-        else:
-            print(story_json.items(), 'xyz')
-            for key, val in story_json.items():
-                if key in ['text', 'title', 'topics', 'image']:
-                    if key == 'topics' and val != []:
-                        for topic in val:
-                            topic_obj = storage._session.query(Topic).where(Topic.name == topic).first()
-                            if topic_obj is not None:
-                                story.topics.append(topic_obj)
-                    else:
-                        setattr(story, key, val)
+@views.route(
+    '/stories/<string:story_id>/',
+    methods=['PUT'],
+    strict_slashes=False
+)
+@auth_guard
+def update_story(story_id=None):
+    story = storage.get(Story, story_id)
+    if story is None:
+        abort(404)
 
-            storage.save()
+    if not auth.authorize(story):
+        return jsonify({"Error": "Permission denied!"}), 403
 
-    return jsonify(story.to_dict()), 200
+    try:
+        story_json = request.get_json()
+        check_for_valid_json(story_json, ['title', 'text'])  # removed check for user_id
 
+    except Exception:
+        return jsonify({"Error": 'not a valid json'}), 400
+
+    else:
+        for key, val in story_json.items():
+            if key in ['text', 'title', 'topics', 'image']:
+                if key == 'topics' and val != []:
+                    for topic in val:
+                        topic_obj = storage._session.query(Topic).where(Topic.name == topic).first()
+                        if topic_obj is not None:
+                            story.topics.append(topic_obj)
+                else:
+                    setattr(story, key, val)
+
+        storage.save()
+    
+    return jsonify(get_story_data(story)), 200
 
 @views.route(
     '/stories/<string:story_id>/',
     methods=['DELETE'],
     strict_slashes=False
 )
-#login_required
+@auth_guard
 def delete_story(story_id=None):
     """ Deletes a story """
     story = storage.get(Story, story_id)
 
     if story is None:
         abort(404)
+
+    if not auth.authorize(story):
+        return jsonify({"Error": "Permission denied!"}), 403
 
     story.delete()
     storage.save()
@@ -217,7 +222,7 @@ def delete_story(story_id=None):
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
+@auth_guard
 def like_or_unlike_story(story_id=None):
     """ Like a story
 
@@ -232,30 +237,127 @@ def like_or_unlike_story(story_id=None):
 
     for like in story.likes:
         # user has already like the story
-        if like.liker.username == current_user.username:
+        if like.liker.username == auth.current_user.username:
             # remove the like
             storage.delete(like)
             storage.save()
-            return jsonify({'status': 'unliked'}), 201
+            return jsonify({'status': 'unliked', 'likes_count': get_story_data(story).get('likes_count')}), 201
 
     # otherwise the user has not like a story
     # like the story
     like = Like(
-        story_id=story_id, user_id=current_user.id
+        story_id=story_id, user_id=auth.current_user.id
     )
     storage.new(like)
     storage.save()
 
-    return jsonify({'status': 'liked'}), 201
+    return jsonify({'status': 'liked', 'likes_count': get_story_data(story).get('likes_count')}), 201
 
 
 @views.route(
-    '/stories/<string:story_id>/<string:user_id>/comments/',
+    '/users/<string:user_id>/follow/',
+    methods=['GET'],
+    strict_slashes=False
+)
+@auth_guard
+def follow_or_unfollow(user_id=None):
+    """ Follow a story
+
+        Attributes:
+            - user_id: id of the user
+
+    """
+
+    user = storage.get(User, user_id)
+    if user is None:
+        abort(404)
+
+    if auth.current_user.is_following(user):
+        auth.current_user.unfollow(user)
+        status = 'unfollowed'
+    else:
+        auth.current_user.follow(user)
+        status = 'follow'
+
+    storage.save()
+
+    return jsonify({status: status}), 201
+
+
+@views.route(
+    '/stories/<string:story_id>/bookmark/',
+    methods=['GET'],
+    strict_slashes=False
+)
+@auth_guard
+def bookmark_or_unbookmark_story(story_id=None):
+    """ Bookmark a story
+
+        Attributes:
+            - story_id: id of the story
+
+    """
+
+    story = storage.get(Story, story_id)
+    if story is None:
+        abort(404)
+
+    for bookmark in story.bookmarks:
+        # user has already bookmark the story
+        if bookmark.bookmarker.username == auth.current_user.username:
+            # remove the bookmark
+            storage.delete(bookmark)
+            storage.save()
+            return jsonify({}), 201
+
+    # otherwise the user has not bookmark a story
+    # bookmark the story
+    bookmark = Bookmark(
+        story_id=story_id, user_id=auth.current_user.id
+    )
+    storage.new(bookmark)
+    storage.save()
+
+    return jsonify({}), 201
+
+
+@views.route(
+    '/comments/<string:comment_id>/like/',
+    methods=['GET'],
+    strict_slashes=False
+)
+@auth_guard
+def like_or_unlike_comment(comment_id=None):
+    """ Like a comment
+
+        Attributes:
+            - comment_id: id of the comment
+
+    """
+    from models.engine import storage
+
+    comment = storage._session.query(Comment).options(joinedload(Comment.commenter)).filter_by(id=comment_id).scalar()
+    if comment is None:
+        abort(404)
+
+    print(comment.is_liked_by(auth.current_user.id), '..................><><><>')
+    if not comment.is_liked_by(auth.current_user.id):  # the user has not like a comment
+        comment.like(auth.current_user.id)  # like the story
+        storage.save()
+        return jsonify({'status': 'liked', 'likes_count': get_comment_data(comment).get('likes_count')}), 201
+    else:  # user has already like the comment
+        comment.unlike(auth.current_user.id)  # unlike it
+        storage.save()
+        return jsonify({'status': 'unliked', 'likes_count': get_comment_data(comment).get('likes_count')}), 201
+
+
+@views.route(
+    '/stories/<string:story_id>/comments/',
     methods=['POST'],
     strict_slashes=False
 )
-#login_required
-def make_comment_on_story(story_id=None, user_id=None):
+@auth_guard
+def make_comment_on_story(story_id=None):
     """ Comment on a story
 
         Attributes:
@@ -264,8 +366,7 @@ def make_comment_on_story(story_id=None, user_id=None):
     """
 
     story = storage.get(Story, story_id)
-    user = storage.get(User, user_id)
-    if story is None or user is None:
+    if story is None or auth.current_user is None:
         abort(404)
 
     try:
@@ -278,28 +379,28 @@ def make_comment_on_story(story_id=None, user_id=None):
         comment = Comment(
             comment=comment_json['comment'],
             story_id=story_id,
-            user_id=user.id
+            user_id=auth.current_user.id
         )
     elif isinstance(comment_json, list):
         comment = Comment(
             comment=comment_json[0],
             story_id=story_id,
-            user_id=user.id
+            user_id=auth.current_user.id
         )
 
     storage.new(comment)
     storage.save()
 
-    return jsonify(comment.to_dict()), 201
+    return jsonify(get_comment_data(comment)), 201
 
 
 @views.route(
-    '/stories/<string:story_id>/<string:user_id>/comments/relevant/',
+    '/stories/<string:story_id>/comments/relevant/',
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
-def get_relevant_comments_for_story(story_id=None, user_id=None):
+@auth_guard
+def get_relevant_comments_for_story(story_id=None):
     """ all comments made on a particular story
 
         Attributes:
@@ -307,13 +408,9 @@ def get_relevant_comments_for_story(story_id=None, user_id=None):
 
     """
 
-    page = request.args.get('page', 1, type=int)
-    per_page = request.args.get('per_page', 10, type=int)
-
     story = storage.get(Story, story_id)
-    user = storage.get(User, user_id)
 
-    if story is None or user is None:
+    if story is None or auth.current_user is None:
         abort(404)
 
     comments_obj = storage._session.execute(story.relevant_comments.options(
@@ -356,28 +453,20 @@ def get_relevant_comments_for_story(story_id=None, user_id=None):
             #  'stories': comments
         #  }
     #  ), 200
-
-    comments = []
-    for comment in comments_obj:
-        temp = comment.to_dict()
-        temp['is_liked_by'] = comment.is_liked_by(user_id)
-        temp['user_is_following_commenter'] = user.is_following(comment.commenter)
-        comments.append(temp)
     
     #  print(comments, '-----------------><>--------------')
 
     return jsonify([
-        comment for comment in comments
+        get_comment_data(comment) for comment in comments_obj
     ])
 
 
 @views.route(
-    '/stories/<string:story_id>/<string:user_id>/comments/newest/',
+    '/stories/<string:story_id>/comments/newest/',
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
-def get_newest_comments_for_story(story_id=None, user_id=None):
+def get_newest_comments_for_story(story_id=None):
     """ all comments made on a particular story
 
         Attributes:
@@ -387,36 +476,26 @@ def get_newest_comments_for_story(story_id=None, user_id=None):
     from models.engine import storage
 
     story = storage.get(Story, story_id)
-    user = storage.get(User, user_id)
 
-    if story is None or user is None:
+    if story is None or auth.current_user is None:
         abort(404)
 
     comments_obj = storage._session.execute(story.newest_comments.options(
         joinedload(Comment.commenter)
     )).scalars().all()
 
-    comments = []
-    for comment in comments_obj:
-        temp = comment.to_dict()
-        temp['is_liked_by'] = comment.is_liked_by(user_id)
-        temp['user_is_following_commenter'] = user.is_following(comment.commenter)
-        comments.append(temp)
-    
-    print(comments, '-----------------><>--------------')
-
     return jsonify([
-        comment for comment in comments
+        get_comment_data(comment) for comment in comments_obj
     ])
 
 
 @views.route(
-    '/stories/<string:story_id>/<string:user_id>/comments/',
+    '/stories/<string:story_id>/comments/',
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
-def get_comments_for_story(story_id=None, user_id=None):
+@auth_guard
+def get_comments_for_story(story_id=None):
     """ all comments made on a particular story
 
         Attributes:
@@ -426,64 +505,17 @@ def get_comments_for_story(story_id=None, user_id=None):
     from models.engine import storage
 
     story = storage.get(Story, story_id)
-    user = storage.get(User, user_id)
 
-    if story is None or user is None:
+    if story is None or auth.current_user is None:
         abort(404)
 
-    comments_obj = storage._session.query(Comment).options(joinedload(Comment.commenter)).filter_by(story_id=story_id).all()
-
-    comments = []
-    for comment in comments_obj:
-        temp = comment.to_dict()
-        temp['is_liked_by'] = comment.is_liked_by(user_id)
-        temp['user_is_following_commenter'] = user.is_following(comment.commenter)
-        comments.append(temp)
-    
-    print(comments, '-----------------><>--------------')
+    comments_obj = storage._session.query(Comment).options(
+        joinedload(Comment.commenter)
+    ).filter_by(story_id=story_id).all()
 
     return jsonify([
-        comment for comment in comments
+        create_uri(get_comment_data(comment), "get_comment_on_story") for comment in comments_obj
     ])
-
-
-
-
-@views.route(
-    '/stories/<string:story_id>/bookmark/',
-    methods=['GET'],
-    strict_slashes=False
-)
-#login_required
-def bookmark_or_unbookmark_story(story_id=None):
-    """ Bookmark a story
-
-        Attributes:
-            - story_id: id of the story
-
-    """
-
-    story = storage.get(Story, story_id)
-    if story is None:
-        abort(404)
-
-    for bookmark in story.bookmark:
-        # user has already bookmark the story
-        if bookmark.bookmarker.username == current_user.username:
-            # remove the bookmark
-            storage.delete(bookmark)
-            storage.save()
-            return jsonify({}), 201
-
-    # otherwise the user has not bookmark a story
-    # bookmark the story
-    bookmark = Bookmark(
-        story_id=story_id, user_id=current_user.id
-    )
-    storage.new(bookmark)
-    storage.save()
-
-    return jsonify({}), 201
 
 
 @views.route(
@@ -491,7 +523,7 @@ def bookmark_or_unbookmark_story(story_id=None):
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
+@auth_guard
 def get_bookmarks_for_story(story_id=None):
     """ all bookmarks made on a particular story
 
@@ -516,7 +548,7 @@ def get_bookmarks_for_story(story_id=None):
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
+@auth_guard
 def get_likes_for_story(story_id=None):
     """ all bookmarks made on a particular story
 
@@ -537,18 +569,16 @@ def get_likes_for_story(story_id=None):
 
 
 @views.route(
-    'liked/<string:story_id>/by/<string:user_id>/',
+    '/users/liked/<string:story_id>/',
     methods=['GET'],
     strict_slashes=False
 )
-#login_required
-def check_like(story_id, user_id):
+@auth_guard
+def check_like(story_id=None):
     """ check if a user has liked a story """
     story = storage.get(Story, story_id)
-    user = storage.get(User, user_id)
 
-    if story or user is None:
+    if not story or auth.current_user is None:
         abort(404)
 
-    return user.liked_story(story.id)
-
+    return jsonify(auth.current_user.liked_story(story.id)), 200
